// src/utils/enums/common.enums.ts
export enum CookiesStorage {
  ACCESS_TOKEN = "accessToken",
  ID_TOKEN = "idToken",
  REFRESH_TOKEN = "refreshToken",
  SESSION_EXPIRE = "sessionExpire",
}

export enum AuthFormVariant {
  REGISTRATION = "register",
  LOGIN = "login",
  RECOVER_PASSWORD = "recoverPassword",
  RESET_PASSWORD = "resetPassword",
  PHONE_VERIFY = "phoneVerification",
  EMAIL_VERIFY = "emailVerification",
}

export enum IconName {
  ADD_PHOTO = "addPhoto",
  ADMIN = "admin",
  APPLE = "apple",
  ARROW = "arrow",
  AMAZON = "amazon",
  AWS = "aws",
  BELL = "bell",
  BOOK = "book",
  BOOKING = "booking",
  CALENDAR = "calendar",
  CARD = "card",
  CHART = "chart",
  CHATS = "chats",
  CLOCK = "clock",
  COPY = "copy",
  DASHBOARD = "dashboard",
  DOWNLOAD = "download",
  DELETE = "delete",
  DELETE_AVATAR = "deleteIcon",
  DELETE_AVATAR_MODAL = "deleteAvatarModal",
  DELETE_USER = "deleteUser",
  EDIT_PHOTO = "editPhoto",
  EMERGENCY = "emergency",
  ERROR = "error",
  EYE = "eye",
  EYE_CLOSED = "eyeClosed",
  EXPAND = "expand",
  FAQ = "faq",
  FINGER_DOWN = "fingerUp",
  FINGER_UP = "fingerUp",
  FILTERS = "filtersIcon",
  GIT = "git",
  GITHUB = "github",
  GOOGLE = "google",
  HANDSHAKE = "handshake",
  LANGUAGE = "language",
  LIGHTNING = "lightning",
  LOG_OUT = "logOut",
  LOGO = "lightLogo",
  LOCK = "lock",
  MAIN = "main",
  MANAGERS = "managers",
  MUI = "mui",
  NOTIFICATIONS = "notifications",
  OFF = "off",
  ORDER = "order",
  REACT = "react",
  REDUX = "redux",
  REQUESTS = "requests",
  SCROLL = "landing_scroll",
  SERVICES = "services",
  SETTINGS = "settings",
  SIDEBAR_BUTTON = "sidebarButton",
  SORT = "sort",
  STATISTICS = "statistics",
  STYLED = "styled",
  TERRAFORM = "terraform",
  TYPESCRIPT = "typescript",
  USERS = "users",
  VITE = "vite",
}

export enum Field {
  CURRENT_PASSWORD = "current_password",
  EMAIL = "email",
  LOGIN = "login",
  PHONE = "phone",
  PASSWORD = "password",
  MESSAGE = "message",
  NEW_PASSWORD = "new_password",
  NUMBER = "number",
}

export enum Locales {
  EN = "en",
  RU = "ru",
}

export const TransactionTypeEnum = {
  CREDIT: "credit",
  PAYMENT: "payment",
};